+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 	# Changelog
 
-## UnRELEASED
+## UNRELEASED
 * Feature: ideas-on-map has a simple view
 * Feature: ideas-on-map details can link to userpage
 * Bugfix: counterbuttons will now work if multiple buttons on exist on a page
 * Bugfix: counterbuttons will no longer show -1
 * Add siteId option to resource overview, allowing for displaying
 * Add ability to link to resource pages with complexer placeholder url (For example #idea:id)
+* Remove deprecated show admin bar
+
 ## 0.7.2 (2020-09-28)
 * Fix for resource admin buttons not appearing
 * Fix for description in resource overview not displaying in gridder